Change quorum threshold type to uint
The type was already being converted to uint
when being passed to SCP, there's no need to
keep it architecture-dependent.

This also allows making the conversion
routines nothrow

+++ source/agora/common/Types.d
@@ -50,7 +50,7 @@ unittest
 public struct QuorumConfig
 {
     /// Threshold of this quorum set
-    public size_t threshold = 1;
+    public uint threshold = 1;
 
     /// List of nodes in this quorum
     public PublicKey[] nodes;
